requirements.txt
beautifulsoup4
lxml
django
djangorestframework
requests
